# [Mono.Android] add a few more _JniMarshal_ built-in delegates