Context: https://github.com/microsoft/dotnet-podcasts

Running the MAUI podcast sample, shows several `_JniMarshal_`
delegates we could list:

    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PP_I': Int32 n_GetItemCount(IntPtr, IntPtr)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PP_I': Int32 n_GetOpacity(IntPtr, IntPtr)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PP_Z': Boolean n_CanScrollVertically(IntPtr, IntPtr)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPI_L': IntPtr n_CreateFragment_I(IntPtr, IntPtr, Int32)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPI_I': Int32 n_GetItemViewType_I(IntPtr, IntPtr, Int32)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPI_I': Int32 n_GetSpanSize_I(IntPtr, IntPtr, Int32)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPI_J': Int64 n_GetItemId_I(IntPtr, IntPtr, Int32)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPLI_L': IntPtr n_OnCreateViewHolder_Landroid_view_ViewGroup_I(IntPtr, IntPtr, IntPtr, Int32)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPLLJ_Z': Boolean n_DrawChild_Landroid_graphics_Canvas_Landroid_view_View_J(IntPtr, IntPtr, IntPtr, IntPtr, Int64)

These look to be the types of APIs that would be common in Android
applications. Listing 7 more of these in `JNINativeWrapper.g.tt` would
cover the above APIs.

A couple more appeared, but these don't seem like they would be as
common:

    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPLF_V': Void n_OnDrawerSlide_Landroid_view_View_F(IntPtr, IntPtr, IntPtr, Single)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPLII_I': Int32 n_OnStartCommand_Landroid_content_Intent_II(IntPtr, IntPtr, IntPtr, Int32, Int32)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPLII_Z': Boolean n_OnError_Landroid_media_MediaPlayer_II(IntPtr, IntPtr, IntPtr, Int32, Int32)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPIII_V': Void n_OnItemRangeMoved_III(IntPtr, IntPtr, Int32, Int32, Int32)
    Falling back to System.Reflection.Emit for delegate type '_JniMarshal_PPLII_V': Void n_OnScrolled_Landroidx_recyclerview_widget_RecyclerView_II(IntPtr, IntPtr, IntPtr, Int32, Int32)

## Results ##

Testing a Release (non-AOT) build of the podcast app on a Pixel 5
device using our `dotnet trace` infrastructure:

https://github.com/xamarin/xamarin-android/blob/main/Documentation/guides/tracing.md

    Before:
    26.88ms (0.56%) mono.android!Android.Runtime.JNINativeWrapper.CreateDelegate(System.Delegate)
     1.20ms (0.02%) mono.android!Android.Runtime.JNINativeWrapper.CreateBuiltInDelegate(System.Delegate,System.Type)

    After:
    17.19ms (0.34%) mono.android!Android.Runtime.JNINativeWrapper.CreateDelegate(System.Delegate)
     1.15ms (0.02%) mono.android!Android.Runtime.JNINativeWrapper.CreateBuiltInDelegate(System.Delegate,System.Type)

I don't think that `CreateBuiltInDelegate` got any faster, but I would
consider it roughly "the same". I think this saves ~9ms in this method.